[release-branch.go1.26] crypto/internal/fips140/drbg: build tag out entropy generation on Wasm
In FIPS-140 mode we currently use a scratch buffer in the BSS
section to generate entropy by measuring jittering for memory
touches. The BSS variable usually doesn't cost much, except on
Wasm, due to the way the linear memory works. FIPS-140 mode is
not supported on Wasm, so this code is not actually needed there.
This CL uses a build tag to exclude it, so we don't need to
include an (unused) 32 MB BSS variable in Wasm binaries.

Updates #78321.
Fixes #78354.

Reviewed-on: https://go-review.googlesource.com/c/go/+/758361
LUCI-TryBot-Result: golang-scoped@luci-project-accounts.iam.gserviceaccount.com <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
(cherry picked from commit e32ec4742e3a9b5a28635e20806331afd587b429)
Reviewed-on: https://go-review.googlesource.com/c/go/+/767320
