+++ src/crypto/internal/fips140/drbg/entropy_wasm.go
@@ -1,0 +1,11 @@
+// Copyright 2026 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+//go:build wasm
+
+package drbg
+
+func readFromEntropy(b []byte) {
+	panic("FIPS-140 entropy generation is not supported on Wasm")
+}
